@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * Created by PhpStorm.
  * User: Inhere
@@ -6,31 +6,21 @@
  * Time: 19:44
  */
 
-namespace Toolkit\Collection;
+namespace PhpPkg\Config;
 
-use InvalidArgumentException;
-use RangeException;
 use RecursiveArrayIterator;
-use stdClass;
-use Toolkit\ArrUtil\Arr;
-use Toolkit\File\Exception\FileNotFoundException;
-use Toolkit\File\File;
-use Toolkit\File\Parse\IniParser;
-use Toolkit\File\Parse\JsonParser;
-use Toolkit\File\Parse\YmlParser;
-use Toolkit\ObjUtil\Obj;
+use Toolkit\Stdlib\Arr;
 use Traversable;
-use function in_array;
 use function is_array;
-use function is_object;
-use function is_string;
 use function serialize;
+use function strpos;
 use function unserialize;
 
 /**
- * Class DataCollector - 数据收集器 (数据存储器 - DataStorage) complex deep
+ * Class Collection
  *
- * @package Toolkit\Collection
+ * @package PhpPkg\Config
+ *
  * 支持 链式的子节点 设置 和 值获取
  * e.g:
  * ```
@@ -41,341 +31,116 @@ use function unserialize;
  *          ]
  *       ]
  * ];
- * $config = new DataCollector();
+ * $config = new Collection();
  * $config->get('foo.bar.yoo')` equals to $data['foo']['bar']['yoo'];
  * ```
- * 简单的数据对象可使用  @see SimpleCollection
- * ```
- * $config = new SimpleCollection($data)
- * $config->get('foo');
- * ```
  */
-class Collection extends SimpleCollection
+class Collection extends \Toolkit\Stdlib\Std\Collection
 {
-    /**
-     * @var array
-     */
-    // protected $files = [];
-
     /**
      * Property separator.
      *
      * @var  string
      */
-    protected $separator = '.';
+    protected string $keyPathSep = '.';
 
     /**
-     * name
+     * set config value by key/path
      *
-     * @var string
-     */
-    protected $name;
-
-    /**
-     * formats
-     *
-     * @var array
-     */
-    protected static $formats = ['json', 'php', 'ini', 'yml'];
-
-    public const FORMAT_JSON = 'json';
-    public const FORMAT_PHP  = 'php';
-    public const FORMAT_INI  = 'ini';
-    public const FORMAT_YML  = 'yml';
-
-    /**
-     * __construct
-     *
-     * @param mixed  $data
-     * @param string $format
-     * @param string $name
-     *
-     * @throws RangeException
-     */
-    public function __construct($data = null, $format = 'php', $name = 'box1')
-    {
-        // Optionally load supplied data.
-        $this->load($data, $format);
-
-        parent::__construct();
-
-        $this->name = $name;
-    }
-
-    /**
-     * @param mixed  $data
-     * @param string $format
-     * @param string $name
-     *
-     * @return static
-     * @throws RangeException
-     */
-    public static function make($data = null, $format = 'php', $name = 'box1')
-    {
-        return new static($data, $format, $name);
-    }
-
-    /**
-     * set config value by path
-     *
-     * @param string $path
-     * @param mixed  $value
+     * @param string $key
+     * @param mixed $value
      *
      * @return mixed
      */
-    public function set($path, $value)
+    public function set(string $key, $value): self
     {
-        Arr::setByPath($this->data, $path, $value, $this->separator);
-
-        return $this;
-    }
-
-    /**
-     * get value by path
-     *
-     * @param string $path
-     * @param string $default
-     *
-     * @return mixed
-     */
-    public function get(string $path, $default = null)
-    {
-        return Arr::getByPath($this->data, $path, $default, $this->separator);
-    }
-
-    public function exists($path): bool
-    {
-        return $this->get($path) !== null;
-    }
-
-    /**
-     * @param string $path
-     *
-     * @return bool
-     */
-    public function has(string $path): bool
-    {
-        return $this->exists($path);
-    }
-
-    public function reset()
-    {
-        $this->data = [];
-
-        return $this;
-    }
-
-    /**
-     * Clear all data.
-     *
-     * @return  static
-     */
-    public function clear()
-    {
-        return $this->reset();
-    }
-
-    /**
-     * @param string $class
-     *
-     * @return mixed
-     */
-    public function toObject(string $class = stdClass::class)
-    {
-        return Arr::toObject($this->data, $class);
-    }
-
-    /**
-     * @return string
-     */
-    public function getSeparator(): string
-    {
-        return $this->separator;
-    }
-
-    /**
-     * @param string $separator
-     */
-    public function setSeparator($separator): void
-    {
-        $this->separator = $separator;
-    }
-
-    /**
-     * @return array
-     */
-    public static function getFormats(): array
-    {
-        return static::$formats;
-    }
-
-    /**
-     * setName
-     *
-     * @param string $value
-     *
-     * @return $this
-     */
-    public function setName($value): self
-    {
-        $this->name = $value;
-
-        return $this;
-    }
-
-    /**
-     * getName
-     *
-     * @return string
-     */
-    public function getName(): string
-    {
-        return $this->name;
-    }
-
-    /**
-     * load
-     *
-     * @param string|array|mixed $data
-     * @param string             $format = 'php'
-     *
-     * @return static
-     * @throws InvalidArgumentException
-     * @throws RangeException
-     */
-    public function load($data, $format = 'php')
-    {
-        if (!$data) {
+        if ($this->keyPathSep && strpos($key, $this->keyPathSep) > 0) {
+            Arr::setByPath($this->data, $key, $value, $this->keyPathSep);
             return $this;
         }
 
-        if (is_string($data) && in_array($format, static::$formats, true)) {
-            switch ($format) {
-                case static::FORMAT_YML:
-                    $this->loadYaml($data);
-                    break;
+        return parent::set($key, $value);
+    }
 
-                case static::FORMAT_JSON:
-                    $this->loadJson($data);
-                    break;
-
-                case static::FORMAT_INI:
-                    $this->loadIni($data);
-                    break;
-
-                case static::FORMAT_PHP:
-                default:
-                    $this->loadArray($data);
-                    break;
-            }
-
-        } elseif (is_array($data) || is_object($data)) {
-            $this->bindData($this->data, $data);
+    /**
+     * @param string $key
+     * @param null $default
+     *
+     * @return mixed
+     */
+    public function get(string $key, $default = null): mixed
+    {
+        if ($this->keyPathSep && strpos($key, $this->keyPathSep) > 0) {
+            return Arr::getByPath($this->data, $key, $default, $this->keyPathSep);
         }
 
+        return parent::get($key, $default);
+    }
+
+    /**
+     * @param string $key
+     *
+     * @return bool
+     */
+    public function exists(string $key): bool
+    {
+        return $this->get($key) !== null;
+    }
+
+    /**
+     * @param string $key
+     *
+     * @return bool
+     */
+    public function has(string $key): bool
+    {
+        return $this->exists($key);
+    }
+
+    /**
+     * @return string
+     */
+    public function getKeyPathSep(): string
+    {
+        return $this->keyPathSep;
+    }
+
+    /**
+     * @param string $keyPathSep
+     */
+    public function setKeyPathSep(string $keyPathSep): void
+    {
+        $this->keyPathSep = $keyPathSep;
+    }
+
+    /**
+     * @param array|Traversable $data
+     *
+     * @return $this
+     */
+    public function load(array|Traversable $data): self
+    {
+        $this->bindData($this->data, $data);
         return $this;
     }
 
     /**
-     * @param        $file
-     * @param string $format
-     *
-     * @return array|mixed
-     * @throws FileNotFoundException
-     */
-    public static function read($file, $format = self::FORMAT_PHP)
-    {
-        return File::load($file, $format);
-    }
-
-    /**
-     * load data form yml file
-     *
-     * @param $data
-     *
-     * @return static
-     */
-    public function loadYaml($data)
-    {
-        return $this->bindData($this->data, static::parseYaml($data));
-    }
-
-    /**
-     * load data form php file or array
-     *
-     * @param array|string $data
-     *
-     * @return static
-     * @throws InvalidArgumentException
-     */
-    public function loadArray($data)
-    {
-        if (is_string($data) && is_file($data)) {
-            $data = require $data;
-        }
-
-        if (!is_array($data)) {
-            throw new InvalidArgumentException('param type error! must is array.');
-        }
-
-        return $this->bindData($this->data, $data);
-    }
-
-    /**
-     * load data form php file or array
-     *
-     * @param mixed $data
-     *
-     * @return static
-     * @throws InvalidArgumentException
-     */
-    public function loadObject($data)
-    {
-        if (!is_object($data)) {
-            throw new InvalidArgumentException('param type error! must is object.');
-        }
-
-        return $this->bindData($this->data, $data);
-    }
-
-    /**
-     * load data form ini file
-     *
-     * @param string $string
-     *
-     * @return static
-     */
-    public function loadIni($string)
-    {
-        return $this->bindData($this->data, self::parseIni($string));
-    }
-
-    /**
-     * load data form json file
-     *
-     * @param $data
-     *
-     * @return Collection
-     */
-    public function loadJson($data): Collection
-    {
-        return $this->bindData($this->data, static::parseJson($data));
-    }
-
-    /**
-     * @param            $parent
-     * @param            $data
-     * @param bool|false $raw
+     * @param array|Traversable $data
      *
      * @return $this
      */
-    protected function bindData(&$parent, $data, $raw = false): self
+    public function loadData(array|Traversable $data): self
     {
-        // Ensure the input data is an array.
-        if (!$raw) {
-            $data = Obj::toArray($data);
-        }
+        $this->bindData($this->data, $data);
+        return $this;
+    }
 
+    /**
+     * @param array $parent
+     * @param array|Traversable $data
+     */
+    protected function bindData(array &$parent, array|Traversable $data): void
+    {
         foreach ($data as $key => $value) {
             if ($value === null) {
                 continue;
@@ -391,8 +156,6 @@ class Collection extends SimpleCollection
                 $parent[$key] = $value;
             }
         }
-
-        return $this;
     }
 
     /**
@@ -418,58 +181,15 @@ class Collection extends SimpleCollection
      *
      * @return  void
      */
-    public function offsetUnset($offset)
+    public function offsetUnset($offset): void
     {
         $this->set($offset, null);
     }
 
     public function __clone()
     {
-        $this->data = unserialize(serialize($this->data), ['allowed_classes' => self::class]);
-    }
-
-    //////
-    ///////////////////////////// helper /////////////////////////
-    //////
-
-    /**
-     * @param               $string
-     * @param bool          $enhancement
-     * @param callable|null $pathHandler
-     * @param string        $fileDir
-     *
-     * @return array
-     */
-    public static function parseIni($string, $enhancement = false, callable $pathHandler = null, $fileDir = ''): array
-    {
-        return IniParser::parse($string, $enhancement, $pathHandler, $fileDir);
-    }
-
-    /**
-     * @param               $data
-     * @param bool          $enhancement
-     * @param callable|null $pathHandler
-     * @param string        $fileDir
-     *
-     * @return array
-     */
-    public static function parseJson($data, $enhancement = false, callable $pathHandler = null, $fileDir = ''): array
-    {
-        return JsonParser::parse($data, $enhancement, $pathHandler, $fileDir);
-    }
-
-    /**
-     * parse YAML
-     *
-     * @param string|bool $data        Waiting for the parse data
-     * @param bool        $enhancement Simple support import other config by tag 'import'. must is bool.
-     * @param callable    $pathHandler When the second param is true, this param is valid.
-     * @param string      $fileDir     When the second param is true, this param is valid.
-     *
-     * @return array
-     */
-    public static function parseYaml($data, $enhancement = false, callable $pathHandler = null, $fileDir = ''): array
-    {
-        return YmlParser::parse($data, $enhancement, $pathHandler, $fileDir);
+        $this->data = unserialize(serialize($this->data), [
+            'allowed_classes' => self::class
+        ]);
     }
 }
